# docs/openapi/baskov-product-api-v1.yaml
openapi: 3.0.3
info:
  title: Baskov Music Product API
  version: 1.0.0
  description: >-
    Authenticated product API and device-session lifecycle used by external Baskov Music clients.
    v1.37 enables bounded personal-library mutations for favorites and shared playlists; Discord remains the source of pairing proof.
servers:
  - url: /
paths:
...
  /api/v1/guilds:
    get:
      operationId: getGuilds
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Discord guilds accessible to the linked identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Guilds'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/home:
    get:
      operationId: getHome
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GuildId'
      responses:
        '200':
          description: Personalized home snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Home'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/mixes:
    get:
      operationId: getMixes
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GuildId'
      responses:
        '200':
          description: Mix cards and continuation state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mixes'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/mixes/{stationSlug}:
    get:
      operationId: getMixDetail
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GuildId'
        - name: stationSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Curated station metadata and seed preview. seedPreview is not a playback queue.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MixDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/search:
    get:
      operationId: searchTracks
      description: Authenticated text search using the same backend-owned YouTube search pipeline as Discord.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GuildId'
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 200
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 5
      responses:
        '200':
          description: Search candidates that can be handed to the existing playback stream endpoint.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Search'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/SearchUnavailable'
  /api/v1/player:
    get:
      operationId: getPlayer
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GuildId'
      responses:
        '200':
          description: Read-only playback snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/library:
    get:
      operationId: getLibrary
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GuildId'
      responses:
        '200':
          description: Personal library snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Library'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/favorites:
    get:
      operationId: getFavorites
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/GuildId' }]
      responses:
        '200': { description: Personal ordered favorites for the linked Discord identity }
    post:
      operationId: addFavorite
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/GuildId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                artist: { type: string }
                title: { type: string }
      responses:
        '200': { description: Updated favorites; adding an existing favorite is idempotent }
        '409': { description: Favorite limit reached }
    delete:
      operationId: clearFavorites
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/GuildId' }]
      responses:
        '200': { description: Empty favorites snapshot }
  /api/v1/favorites/{position}:
    delete:
      operationId: removeFavorite
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/GuildId'
        - { name: position, in: path, required: true, schema: { type: integer, minimum: 1 } }
      responses:
        '200': { description: Updated favorites }
        '404': { description: Favorite position not found }
  /api/v1/playlists:
    get:
      operationId: getPlaylists
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/GuildId' }]
      responses:
        '200':
          description: Guild shared playlists visible to the linked Discord identity
    post:
      operationId: createPlaylist
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/GuildId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        '200': { description: Created shared playlist }
        '409': { description: Playlist already exists or guild limit reached }
  /api/v1/playlists/{name}:
    get:
      operationId: getPlaylist
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/GuildId'
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Shared playlist detail }
    delete:
      operationId: deletePlaylist
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/GuildId'
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Deleted playlist snapshot }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/playlists/{name}/tracks:
    post:
      operationId: addPlaylistTrack
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/GuildId'
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Updated playlist detail }
  /api/v1/playlists/{name}/tracks/{position}:
    delete:
      operationId: removePlaylistTrack
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/GuildId'
        - { name: name, in: path, required: true, schema: { type: string } }
        - { name: position, in: path, required: true, schema: { type: integer, minimum: 1 } }
      responses:
        '200': { description: Updated playlist detail }
  /api/v1/playlists/{name}/move:
    post:
      operationId: movePlaylistTrack
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/GuildId'
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Updated playlist detail }
  /api/v1/playlists/{name}/rename:
    post:
      operationId: renamePlaylist
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/GuildId'
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Renamed playlist detail }
  /api/v1/playback/stream:
    get:
      operationId: streamMobilePlayback
      description: >-
        Authenticated foreground Ogg/Opus stream for one provider-neutral logical track.
        The backend owns provider selection and source loading; clients send artist/title only.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GuildId'
        - name: artist
          in: query
          required: true
          schema:
            type: string
            maxLength: 180
        - name: title
          in: query
          required: true
          schema:
            type: string
            maxLength: 180
      responses:
        '200':
          description: Streaming Ogg container with Opus audio packets
          headers:
            X-Baskov-Playback-Duration-Millis:
              schema:
                type: integer
                format: int64
          content:
            audio/ogg:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/PlaybackUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: BaskovAccessToken
  parameters:
    GuildId:
      name: guildId
      in: query
      required: true
      description: Discord guild snowflake in decimal form.
      schema:
        type: string # Discord snowflake
        pattern: '^[1-9][0-9]{0,19}$'
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing, invalid, expired or revoked credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Linked Discord identity cannot access requested guild
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    SearchUnavailable:
      description: Search source temporarily unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    PlaybackUnavailable:
      description: No healthy provider/track source can prepare a mobile foreground stream
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Guilds:
      type: object
      required: [userId, guilds]
      properties:
        userId:
          type: string
          format: uuid
        guilds:
          type: array
          items:
            $ref: '#/components/schemas/Guild'
...
    Home:
      type: object
      required: [guildId, userId, date, today, forYou, themes, library, recent, taste]
      properties:
        guildId:
          type: string # Discord snowflake
        userId:
          type: string
          format: uuid
        date:
          type: string
          format: date
        continuation:
          allOf:
            - $ref: '#/components/schemas/Continuation'
          nullable: true
        today:
          type: array
          items:
            $ref: '#/components/schemas/Mix'
        forYou:
          type: array
          items:
            $ref: '#/components/schemas/Mix'
        themes:
          type: array
          items:
            $ref: '#/components/schemas/Theme'
        library:
          $ref: '#/components/schemas/Library'
        recent:
          type: array
          items:
            $ref: '#/components/schemas/Track'
        taste:
          $ref: '#/components/schemas/Taste'
    Mixes:
      type: object
      required: [guildId, userId, date, today, forYou, themes]
      properties:
        guildId:
          type: string # Discord snowflake
        userId:
          type: string
          format: uuid
        date:
          type: string
          format: date
        continuation:
          allOf:
            - $ref: '#/components/schemas/Continuation'
          nullable: true
        today:
          type: array
          items:
            $ref: '#/components/schemas/Mix'
        forYou:
          type: array
          items:
            $ref: '#/components/schemas/Mix'
        themes:
          type: array
          items:
            $ref: '#/components/schemas/Theme'
    Player:
      type: object
      required: [guildId, sessionActive, playing, paused, volume, repeatMode, queueSize, positionMillis, durationMillis, radio]
      properties:
        guildId:
          type: string # Discord snowflake
        sessionActive:
          type: boolean
        playing:
          type: boolean
        paused:
          type: boolean
        volume:
          type: integer
        repeatMode:
          type: string
        queueSize:
          type: integer
        positionMillis:
          type: integer
          format: int64
        durationMillis:
          type: integer
          format: int64
        current:
          allOf:
            - $ref: '#/components/schemas/Track'
          nullable: true
        radio:
          $ref: '#/components/schemas/Radio'
    Library:
      type: object
      required: [guildId, userId, favorites, personalHistory, recent, favoriteTracks, historyTracks]
      properties:
        guildId:
          type: string # Discord snowflake
        userId:
          type: string
          format: uuid
        favorites:
          type: integer
        personalHistory:
          type: integer
        recent:
          type: array
          items:
            $ref: '#/components/schemas/Track'
        favoriteTracks:
          type: array
          items:
            $ref: '#/components/schemas/Track'
        historyTracks:
          type: array
          items:
            $ref: '#/components/schemas/Track'
    Search:
      type: object
      required: [guildId, userId, query, tracks]
      properties:
        guildId:
          type: string # Discord snowflake
        userId:
          type: string
          format: uuid
        query:
          type: string
        tracks:
          type: array
          items:
            $ref: '#/components/schemas/Track'
    MixDetail:
      type: object
      required: [guildId, userId, stationSlug, label, description, available, daily, seedPreview]
      properties:
        guildId:
          type: string
        userId:
          type: string
          format: uuid
        stationSlug:
          type: string
        label:
          type: string
        description:
          type: string
        available:
          type: boolean
        daily:
          type: boolean
        seedPreview:
          type: array
          description: Read-only station seed preview; not a predicted playback queue.
          items:
            $ref: '#/components/schemas/Track'
...